requirements.txt
Django==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.6
python-dotenv==1.0.0
django-cors-headers==4.0.0
drf-yasg==1.21.5
pytest==7.3.1
pytest-django==4.5.2
django-filter==23.2
Pillow==10.0.0